pyproject.toml
[project]
name = "agent-playground"
version = "0.1.0"
description = "Agent Playground is a simple sandbox for building and testing AI agents"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "google-genai>=1.69.0",
    "httpx>=0.28.1",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "restrictedpython>=8.1",
]
